imageclassify/app/requirements.txt
torch
torchvision
flask
flask-cors
numpy
pillow
